requirements.txt
crewai
crewai-tools
langchain
langchain_google_genai
python-dotenv
fastapi
uvicorn
pandas
plotly
reportlab
streamlit
tenacity
google-generativeai
PyPDF2
pdfplumber
